+++ components/NavigationDrawer/src/private/MDCBottomDrawerContainerViewController.h
@@ -13,8 +13,44 @@
 // limitations under the License.
 
 #import <UIKit/UIKit.h>
+#import "MDCBottomDrawerState.h"
 
+@class MDCBottomDrawerContainerViewController;
 @protocol MDCBottomDrawerHeader;
+@protocol MDCBottomDrawerContainerViewControllerDelegate;
+
+/**
+ Delegate for MDCBottomDrawerContainerViewController.
+ */
+@protocol MDCBottomDrawerContainerViewControllerDelegate <NSObject>
+/**
+ This method is called when the bottom drawer will change its presented state to one of the
+ MDCBottomDrawerState states.
+
+ @param containerViewController the container view controller of the bottom drawer.
+ @param drawerState the drawer's state.
+ */
+- (void)bottomDrawerContainerViewControllerWillChangeState:
+            (nonnull MDCBottomDrawerContainerViewController *)containerViewController
+                                               drawerState:(MDCBottomDrawerState)drawerState;
+
+/**
+ This method is called when the drawer is scrolled/dragged and provides a transition ratio value
+ between 0-100% (0-1) that indicates the percentage in which the drawer is close to reaching the end
+ of its scrolling. If the drawer is about to reach fullscreen, its percentage moves between 0-100%
+ as it starts covering the safe area and status bar. If the drawer doesn't reach full screen, it
+ moves between 0-100% as it reaches 20 points away from being fully expanded.
+ - 1 or 100% indicates the drawer has reached the end of its scrolling upwards to reveal content.
+ - 0 or 0% indicates that there is more scrolling to be done for the drawer to either present
+ more content or to transition to full screen.
+
+ @param containerViewController the container view controller of the bottom drawer.
+ @param transitionRatio The transition ratio betwen 0-100% (0-1).
+ */
+- (void)bottomDrawerContainerViewControllerTopTransitionRatio:
+            (nonnull MDCBottomDrawerContainerViewController *)containerViewController
+                                              transitionRatio:(CGFloat)transitionRatio;
+@end
 
 /**
  View controller for containing a Google Material bottom drawer. Used internally only.
@@ -60,5 +96,15 @@
 
 // Whether the drawer is currently animating its presentation.
 @property(nonatomic) BOOL animatingPresentation;
+
+/**
+ Delegate to tell the presentation controller when the drawer will change state.
+ */
+@property(nonatomic, weak, nullable) id<MDCBottomDrawerContainerViewControllerDelegate> delegate;
+
+/**
+ The current state of the bottom drawer.
+ */
+@property(nonatomic, readonly) MDCBottomDrawerState drawerState;
 
 @end