+++ components/NavigationDrawer/src/private/MDCBottomDrawerContainerViewController.m
@@ -33,7 +33,6 @@ static const CGFloat kEpsilon = 0.001f;
 // smooth.
 static const CGFloat kScrollViewBufferForPerformance = 20.f;
 static const CGFloat kDragVelocityThresholdForHidingDrawer = -2.f;
-static const CGFloat kDefaultHeaderCornerRadius = 8.f;
 static NSString *const kContentOffsetKeyPath = @"contentOffset";
 
 static UIColor *DrawerShadowColor(void) {
@@ -137,6 +136,9 @@ static UIColor *DrawerShadowColor(void) {
 // The top header bottom shadow layer.
 @property(nonatomic) MDCShadowLayer *headerShadowLayer;
 
+// The current bottom drawer state.
+@property(nonatomic) MDCBottomDrawerState drawerState;
+
 @end
 
 @implementation MDCBottomDrawerContainerViewController {
@@ -146,7 +148,6 @@ static UIColor *DrawerShadowColor(void) {
   CGFloat _contentHeightSurplus;
   CGFloat _addedContentHeight;
   CGFloat _contentVCPreferredContentSizeHeightCached;
-  MDCBottomDrawerHeaderMask *_maskLayer;
 }
 
 - (instancetype)initWithOriginalPresentingViewController:
@@ -159,9 +160,7 @@ static UIColor *DrawerShadowColor(void) {
     _contentHeightSurplus = NSNotFound;
     _addedContentHeight = NSNotFound;
     _trackingScrollView = trackingScrollView;
-    _maskLayer =
-        [[MDCBottomDrawerHeaderMask alloc] initWithMaximumCornerRadius:kDefaultHeaderCornerRadius
-                                                   minimumCornerRadius:0];
+    _drawerState = MDCBottomDrawerStateCollapsed;
   }
   return self;
 }
@@ -319,6 +318,22 @@ static UIColor *DrawerShadowColor(void) {
   [self.scrollView removeObserver:self forKeyPath:kContentOffsetKeyPath];
 }
 
+- (void)setDrawerState:(MDCBottomDrawerState)drawerState {
+  if (drawerState != _drawerState) {
+    _drawerState = drawerState;
+    [self.delegate bottomDrawerContainerViewControllerWillChangeState:self drawerState:drawerState];
+  }
+}
+
+- (void)updateDrawerState:(CGFloat)transitionPercentage {
+  if (transitionPercentage >= 1.f - kEpsilon) {
+    self.drawerState = self.contentReachesFullscreen ? MDCBottomDrawerStateFullScreen
+                                                     : MDCBottomDrawerStateExpanded;
+  } else {
+    self.drawerState = MDCBottomDrawerStateCollapsed;
+  }
+}
+
 #pragma mark UIViewController
 
 - (void)viewDidLoad {
@@ -367,14 +382,6 @@ static UIColor *DrawerShadowColor(void) {
     scrollViewFrame.origin.y = -self.topHeaderHeight;
     self.scrollView.frame = scrollViewFrame;
   } else {
-    if (self.contentHeightSurplus > 0) {
-      if (self.topHeaderHeight != 0.f) {
-        _maskLayer.view = self.headerViewController.view;
-      } else {
-        _maskLayer.view = self.contentViewController.view;
-      }
-      [_maskLayer applyMask];
-    }
     CGRect scrollViewFrame = self.presentingViewBounds;
     if (self.animatingPresentation) {
       CGFloat heightSurplusForSpringAnimationOvershooting =
@@ -475,7 +482,9 @@ static UIColor *DrawerShadowColor(void) {
                                         distance:self.headerAnimationDistance];
   CGFloat headerTransitionToTop =
       contentOffset.y >= self.transitionCompleteContentOffset ? 1.f : transitionPercentage;
-  [_maskLayer animateWithPercentage:1.f - transitionPercentage];
+  [self.delegate bottomDrawerContainerViewControllerTopTransitionRatio:self
+                                                       transitionRatio:transitionPercentage];
+  [self updateDrawerState:transitionPercentage];
   self.currentlyFullscreen = self.contentReachesFullscreen && headerTransitionToTop >= 1.f;
   CGFloat fullscreenHeaderHeight =
       self.contentReachesFullscreen ? self.topHeaderHeight : [self contentHeaderHeight];
@@ -669,6 +678,13 @@ static UIColor *DrawerShadowColor(void) {
 
   CGFloat scrollingDistance = _contentHeaderTopInset + contentHeaderHeight + contentHeight;
   _contentHeightSurplus = scrollingDistance - containerHeight;
+  if ([self shouldPresentFullScreen]) {
+    self.drawerState = MDCBottomDrawerStateFullScreen;
+  } else if (_contentHeightSurplus <= 0) {
+    self.drawerState = MDCBottomDrawerStateExpanded;
+  } else {
+    self.drawerState = MDCBottomDrawerStateCollapsed;
+  }
   if (addedContentHeight < kEpsilon && (_contentHeaderTopInset > _contentHeightSurplus) &&
       (_contentHeaderTopInset - _contentHeightSurplus < self.addedContentHeightThreshold)) {
     CGFloat addedContentheight = _contentHeaderTopInset - _contentHeightSurplus;
@@ -684,7 +700,7 @@ static UIColor *DrawerShadowColor(void) {
 }
 
 - (CGFloat)midAnimationScrollToPositionForOffset:(CGPoint)targetContentOffset {
-  if (!self.contentScrollsToReveal) {
+  if (!self.contentScrollsToReveal || !self.contentReachesFullscreen) {
     return NSNotFound;
   }
 
@@ -693,7 +709,6 @@ static UIColor *DrawerShadowColor(void) {
       [self transitionPercentageForContentOffset:targetContentOffset
                                           offset:0
                                         distance:headerAnimationDistance];
-  [_maskLayer animateWithPercentage:1.f - headerTransitionToTop];
   if (headerTransitionToTop >= kEpsilon && headerTransitionToTop < 1.f) {
     CGFloat contentHeaderFullyCoversTopHeaderContentOffset = self.transitionCompleteContentOffset;
     CGFloat contentHeaderReachesTopHeaderContentOffset =